src/requirements.txt
flask
python-ldap
gunicorn
